Changed duration of "Destello"
